// src/types/feature/explorer/connectionsTreeRow.types.ts
export type ConnectionsTreeRowKind =
  | "keyspaces_folder"
  | "keyspace"
  | "tables_folder"
  | "table"
  | "columns_folder"
  | "column"
  | "indexes_folder"
  | "index"
  | "mv_folder"
  | "mv"
  | "udt_folder"
  | "udt"
  | "functions_folder"
  | "function"
  | "aggregates_folder"
  | "aggregate"
  | "roles_folder"
  | "role"
  | "permissions_folder"
  | "permission"
  | "system_keyspace_folder";